+++ CHANGELOG.md
@@ -9,6 +9,12 @@ develop branch)  won't be reflected in this file.
 
 ## [Unreleased]
 
+### Added
+- bumpversion support
+
+
+### Deprecated
+- `taurus.Release.version_info` and `taurus.Release.revision` variables
 
 ## [4.0.1] - 2016-07-19
 Jul16 milestone.